typeahead: Add recent_senders module.
In recent_senders module, added a data structure to hold timestamps of
users' latest message in a topic. Also added a function to compare 2 users
based on above timestamp. Added a function to process messages for the
data structure and a call in add_message_metadata. Also added node tests
for insertion of data into recent_senders.senders.