org-settings: Add UI for notifications stream.
Added a dropdown in the organization settings page with a search-box and
required styles. Added a method to populate the dropdown using
list_rendering.js. Also altered response to the event of deletion of the
notifications stream on the frontend. On form submission, selected
stream id is sent to /json/realm.

Fixes: #3708